Add scripts for getting and generating the OpenStack RAG
This patch is adding two new scripts:

* The get_openstack_plaintext_docs.sh

This scripts downloads and builds the OpenStack documentation for the
core projects in plain text format.

Once it runs it will output a folder called "openstack-docs-plaintext"
in the local directory containing the documentation in plain text.

Usage:
  ./scripts/get_openstack_plaintext_docs.sh 2024.2

2. generate_embeddings_openstack.py

Just like the OpenShift counterpart, this script will generate a RAG
database based on the OpenStack documentation.

Usage:
  python3 scripts/generate_embeddings_openstack.py -o ./vector_db/
-f openstack-docs-plaintext/ -md embeddings_model/ -mn
sentence-transformers/all-mpnet-base-v2 -i os-docs-2024_2

This patch also separated the common code for the generate_embeddings*
scripts into a common file called "common_embedding" and later we can
modify the generate_embeddings.py script to also use it and share most
of the code between the two scripts.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>