Fix the OpenStack RAG generation metadata
During some refactors we stopped using the OS_DOCS_ROOT_URL for our
metadata, instead, it was pointing to the filepath in the filesystem.
This led to all URLs being marked as unreacheable.

Also fix the help message of a config option which no longer defaults to
the number of CPUs.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>