Install lightspeed-rag-content in the base image (#30)
Modifies the Containerfile.base so that lightspeed-rag-content package is installed inside of the image.
Right now the building of the vector database inside of the image does not
work for the OpenShift example because the package is not present.

Including the lightspeed-rag-content package in the base image will
make it easier for other repositories (e.g.,
openstack-lightspeed/rag-content [1]) to use.

[1] https://github.com/openstack-lightspeed/rag-content